+++ tests/ImageProcessing/Commands/MoveToS3Test.php
@@ -16,7 +16,7 @@
  * @noinspection PhpUnhandledExceptionInspection
  */
 
-namespace Tests\Feature_v2\Commands;
+namespace Tests\ImageProcessing\Commands;
 
 use Tests\Feature_v2\Base\BaseApiWithDataTest;
